+++ src/domains/filter/apis/mockSearchApi.ts
@@ -1,6 +1,6 @@
 // src/domains/filter/services/mockSearchApi.ts
 
-// 🎸 임시 데이터
+// 임시 데이터
 const mockResults = [
   { id: 1, name: "Fender Stratocaster", price: 1200000 },
   { id: 2, name: "Yamaha Drum Set", price: 800000 },
@@ -8,7 +8,7 @@ const mockResults = [
   { id: 4, name: "Gibson Les Paul", price: 1700000 },
 ];
 
-// 🧪 백엔드 API 대기 중: 가짜 검색 함수
+// 백엔드 API 대기 중: 가짜 검색 함수
 export async function searchProductsMock(keyword: string) {
   // API 지연 시뮬레이션
   await new Promise((resolve) => setTimeout(resolve, 400));
